requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
xgboost>=2.0
tensorflow-cpu==2.15.*
matplotlib>=3.7
pydantic>=2.6
fastapi>=0.110
uvicorn[standard]>=0.27
pyyaml>=6.0